# docs/openapi_flowers.yaml
openapi: 3.0.0
info: 
  title: "Цветы"
  version: 0.0.1
  description: |
    MVP из говна и палок 

    Для использования бота нужно сделать отдельную авторизацию с Oauth телеги. Пока стандартный шаблон.
    
    Last update: 00:00 27.02.24 by NF
servers:
  - url: http://localhost:8453/api/v1/
    description: API version 1
paths:
...
  /order/myActiveOrders:
    get:
      summary: Информация об активных заказах пользователя (+)
      tags:
        - User Order
      security:
        - bearerAuth: []
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderStatus"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /order/orderInfo:
    get:
      summary: Информация о заказе по идентификатору (+)
      tags:
        - User Order
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: orderId
          required: true
          schema:
            $ref: "#/components/schemas/OrderId"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStatus"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /order/createOrder:
    post:
      summary: Создание заказа (+)
      tags:
        - User Order
      security:
        - bearerAuth: []
      description: |
        Пока опускаем момент с оплатой, интегрируем её потом. Скорее всего она не повлияет на данный эндпоинт
        
        TODO: добавить возможность отложенной доставки. До тех пор пользователи должны указывать подобное пожелание в комментариях
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Geo:
                  $ref: "#/components/schemas/Adress"
                FirstName:
                  $ref: "#/components/schemas/FirstName"
                SecondName:
                  $ref: "#/components/schemas/SecondName"
                Comment:
                  $ref: "#/components/schemas/Comment"
                PhoneNumber:
                  $ref: "#/components/schemas/PhoneNumber"
                ProductIdArray:
                  $ref: "#/components/schemas/ProductIdArray"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultPing"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /order/cancelOrder:
    post:
      summary: Отмена заказа (+)
      description: |
        Возможность отменить заказ до какого-либо момента
      tags:
        - User Order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
         application/json:
          schema:
            type: object
            properties:
              orderId:
                $ref: "#/components/schemas/OrderId"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DefaultPing"
        #403:
        #  description: Responce with data
        #  content:
        #    application/json:
        #      schema:
        #        $ref: "#/components/schemas/DefaultPing"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /catalog/getCatalog:
    get:
      summary: Каталог товаров (+-)
      tags:
        - Catalog
      parameters:
        - in: query
          name: start
          required: true
          schema:
            type: integer
            description: номер карточки товара в поиске
            example: 0
        - in: query
          name: count
          required: false
          schema:
            type: integer
            description: число карточек в выдаче
            example: 20
        - in: query
          name: category
          required: false
          schema:
            type: string
            description: категория товара
            example: Небритые фиалки
        - in: query
          name: sort
          required: false
          schema:
            type: string
            description: |
             Один из вариантов: default, price_upscending, price_descending
            example: default
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItemsArray"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /catalog/search:
    get:
      summary: Каталог товаров (+)
      tags:
        - Catalog
      parameters:
        - in: query
          name: request
          required: true
          schema:
            type: string
            description: поисковой запрос
            example: Букет с либерахами
        - in: query
          name: start
          required: true
          schema:
            type: integer
            description: номер карточки товара в поиске
            example: 0
        - in: query
          name: count
          required: false
          schema:
            type: integer
            description: число карточек в выдаче
            example: 20
        - in: query
          name: category
          required: false
          schema:
            type: string
            description: категория товара
            example: Небритые фиалки
        - in: query
          name: sort
          required: false
          schema:
            type: string
            description: |
             Один из вариантов: default, price_upscending, price_descending
            example: default
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItemsArray"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /catalog/getCatalogItemDetails:
    get:
      summary: Получение деталей о товаре (+)
      tags:
        - Catalog
      parameters:
        - in: query
          name: id
          required: true
          schema:
            $ref: "#/components/schemas/ProductId"
      responses:
        200:
          description: Responce with data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogItemDetails"
        default:
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    basicAuth:
      type: http
      scheme: basic
  schemas:
    FirstName:
      type: string
      description: Имя
      maxLength: 30
      pattern: "[a-zA-Z]+"
      example: Макака
    SecondName:
      type: string
      description: Фамилия
      maxLength: 50
      pattern: "[a-zA-Z]+"
      example: Тупая
    PhoneNumber:
      type: string
      description: Номер телефона
      maxLength: 16
      pattern: "/+[0-9]{11,15}"
      example: +78005553535
...
    OrderId:
      type: integer
      description: Номер заказа
      example: 1
    ProductId:
      type: integer
      description: Номер товара
      example: 1
    Comment:
      type: string
      description: Комментарий от заказчика
      maxLength: 300
      example: Домофон не работает, позвоните на номер телефона, указанный в заказе
    Adress:
      type: object
      description: Адрес гоя
      properties:
        Country:
          type: string
          description: "Country of costumer"
          example: "Russia"
        City:
          type: string
          description: "City/town or etc of costumer"
          example: "St. Petersburg"
        Street:
          type: string
          description: "Street or etc of costumer"
          example: "Lensoveta street"
        Building:
          type: string
          description: "Home number (with additional data like letter)"
          example: "23"
        Flat:
          type: string
          description: "Flat/Office number"
          example: "147"
...
    ProductIdArray:
      type: array
      description: массив номеров заказанных продуктов
      items:
        $ref: "#/components/schemas/ProductId"
...
    OrderStatus:
      type: object
      required:
        - orderId
        - orderStatus
        - orderCreatedTime
        - adress
        - comment
        - customerPhone
        - customerFirstName
        - customerSecondName
        - productIdArray
      properties:
        orderId:
          $ref: "#/components/schemas/OrderId"
        adress:
          $ref: "#/components/schemas/Adress"
        orderStatus:
          type: string
          description: Статус заказа # maybe we must use enum to create pretty bar (or add smth like stage number)
          example: Обрабатывается...
        orderCreatedTime:
          type: integer
          description: timestamp
          example: 1903093
        customerPhone:
          $ref: "#/components/schemas/PhoneNumber"
        customerFirstName:
          $ref: "#/components/schemas/FirstName"
        customerSecondName:
          $ref: "#/components/schemas/SecondName"
        comment:
          $ref: "#/components/schemas/Comment"  
        productIdArray:
          $ref: "#/components/schemas/ProductIdArray" 
...
    CatalogItemTitle:
      type: string
      description: "Краткое описание букета"
      example: "Букет из чего-то там"
    CatalogItemAuthor:
      type: string
      description: "Название магазина"
      example: "ООО Рога и копыта"
    CatalogItemImage:
      type: string
      description: "Ссылка на картинку с букетом"
      example: "https://example.com/test.png"
    CatalogItemCost:
      type: object
      properties:
        currency: # задел на будуюущее для расширения
          type: string
          description: "Валюта"
          example: RUB
        costNum:
          type: integer
          description: Число тубриков
          example: 30
    CatalogItemReviewsInfo:
      type: object
      properties:
        rank:
          type: number
          description: "Рейтинг товара c точностью до 1 знака после запятой с округлением вверх (4.001=4.1)"
          example: 4.8
        reviewsCount:
          type: integer
          description: Число отзыврв
          example: 30
    CatalogItem:
      type: object
      description: Элемент каталога
      properties:
        title:
          $ref: "#/components/schemas/CatalogItemTitle"
        author:
          $ref: "#/components/schemas/CatalogItemAuthor"
        image:
          $ref: "#/components/schemas/CatalogItemImage"
        productId:
          $ref: "#/components/schemas/ProductId"
        cost:
          $ref: "#/components/schemas/CatalogItemCost"
        reviews:
          $ref: "#/components/schemas/CatalogItemReviewsInfo"
    CatalogItemsArray:
      type: array
      description: Массив элементов каталога
      items:
        $ref: "#/components/schemas/CatalogItem"
    CatalogItemDescription:
      type: string
      description: Описание элемента каталога
      example: |
        Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book.
    CatalogItemBoughtTimesCounter:
      type: integer
      description: Сколько раз был куплен товар
      example: 100
    CatalogItemDetails:
      type: object
      description: Детальное описание элемента каталога
      properties:
        title:
          $ref: "#/components/schemas/CatalogItemTitle"
        author:
          $ref: "#/components/schemas/CatalogItemAuthor"
        titleImage:
          $ref: "#/components/schemas/CatalogItemImage"
        additionalImages:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItemImage"
        productId:
          $ref: "#/components/schemas/ProductId"
        cost:
          $ref: "#/components/schemas/CatalogItemCost"
        reviews:
          $ref: "#/components/schemas/CatalogItemReviewsInfo"
        description:
          $ref: "#/components/schemas/CatalogItemDescription"
        boughtTimesCounter:
          $ref: "#/components/schemas/CatalogItemBoughtTimesCounter"
...
    DefaultPing:
      type: object
      properties:
        data:
          type: string
          description: "DO NOT USE THIS FIELD! It's placeholder and can be removed anytime/"
          example: "gooooool!"
...
    Error:
      type: object
      required:
        - description
      properties:
        code:
          type: integer
          description: "Code of error (optional! it's better to get status code from responce directly!)"
          example: 500
        description:
          type: string
          description: "Description of error"
          example: description of error